# implementation_examples/openapi.yaml
openapi: 3.1.0
...
info:
  title: Oolix API
  version: 0.2.0
  description: |
    Privacy-safe partner media activation control plane.

    **The rule every endpoint serves.** Oolix is the control plane; Data
    Partners remain the customer-data owners (§1). No request or response in
    this contract carries a `partner_user_id`, a segment membership list or an
    exact audience count. Reach is always a bucket (§72) and attribution is
    always an opaque token resolved server-side by SHA-256 (§90).
  contact:
    name: Oolix engineering
  license:
    name: Oolix proprietary
    identifier: LicenseRef-Oolix-Proprietary
servers:
  - url: http://localhost:4000
    description: Local development (docker compose stack)
  - url: '{scheme}://{host}'
    description: |
      Deployed environments. The host is supplied per environment rather than
      hard-coded, so a generated client can never default to someone else's
      domain.
    variables:
      scheme: { default: https, enum: [https] }
      host: { default: localhost:4000 }
...
security:
  - oidcBearer: []
paths:
...
  /v1/attribution/click/{token}:
    post:
      operationId: recordClick
      tags: [Attribution]
      summary: Record a click against an opaque token
      description: |
        §14, §90. The response is **identical** whether or not the token is
        recognised -- a differing status would let a caller probe for valid
        tokens.
      security: []
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Attribution outcome
          content:
            application/json:
              schema:
                type: object
                properties:
                  attributed: { type: boolean }
                  activation_id: { type: string, format: uuid }

        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/leads/events:
    post:
      operationId: reportLeadEvent
      parameters:
        - $ref: '#/components/parameters/OrgId'
      tags: [Attribution]
      summary: Report a CRM lead event
      description: |
        §71. Authenticated by the Buyer's **CRM API key**, not an OIDC token:
        the caller is a machine in the Buyer's own stack.

        Transitions are forward-only (`RECEIVED -> VALID -> QUALIFIED ->
        CONVERTED`, or `-> REJECTED`). Idempotent on
        `(buyer_org_id, crm_event_id)` so a retry storm cannot inflate the lead
        count -- and therefore the payout.
      security:
        - crmApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LeadEvent' }
      responses:
        '201': { description: 'Event accepted or recognised as a duplicate' }
        '401': { description: Invalid or missing CRM API key }

        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/leads/events/batch:
    post:
      operationId: reportLeadEventBatch
      parameters:
        - $ref: '#/components/parameters/OrgId'
      tags: [Attribution]
      summary: Report many CRM lead events
      description: |
        Each event is independent: one invalid event does not reject the batch,
        because a CRM retrying the whole batch would stall every other lead.
      security:
        - crmApiKey: []
      responses:
        '201': { description: Per-event results }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/conversions/events:
    post:
      operationId: reportConversionEvent
      parameters:
        - $ref: '#/components/parameters/OrgId'
      tags: [Attribution]
      summary: Report a conversion
      description: |
        §71: a conversion is a lead-state transition, not a separate object --
        a click token maps to ONE logical lead moving through states.
      security:
        - crmApiKey: []
      responses:
        '201': { description: Conversion recorded }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/buyer/crm-keys:
    post:
      operationId: rotateCrmApiKey
      parameters:
        - $ref: '#/components/parameters/OrgId'
      tags: [Attribution]
      summary: Issue or rotate the Buyer's CRM API key
      description: |
        §71: keys are hashed at rest and can be rotated. The plaintext is
        returned **once**.
      responses:
        '201':
          description: New key (shown once)
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_key: { type: string }
                  warning: { type: string }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  # ---------------------------------------------------------------------------
...
  /v1/billing/payouts/calculate:
    post:
      operationId: calculatePayout
      tags: [Billing]
      summary: Calculate a payout and write immutable FinancialEvents
      description: |
        §50, §102. Writes `OUTCOME_ACCRUAL`, `PLATFORM_FEE` and
        `PARTNER_PAYOUT_BASIS` events. Re-running is idempotent on
        `(activation, type, source_ref)` and does not duplicate the accrual.
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CalculatePayout' }
      responses:
        '201':
          description: Settlement, with the immutable inputs it was derived from
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Settlement' }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/billing/payouts/{id}/reproduce:
    get:
      operationId: reproduceSettlement
      tags: [Billing]
      summary: Re-derive a settlement from its immutable inputs
      description: |
        §50 rule 4: "Every financial calculation should be reproducible from
        immutable settlement inputs." A Partner does not have to take the
        figure on trust -- it can be recomputed and compared.
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Reproduction result
          content:
            application/json:
              schema:
                type: object
                properties:
                  reproducible: { type: boolean }
                  original: { type: object, additionalProperties: true }
                  recomputed: { type: object, additionalProperties: true }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/billing/payouts:
    get:
      operationId: listPayouts
      parameters:
        - $ref: '#/components/parameters/OrgId'
      tags: [Billing]
      summary: The Partner's payouts, with every contributing event
      responses:
        '200': { description: Payouts }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/billing/payouts/{id}/review:
    post:
      operationId: reviewPayout
      tags: [Billing]
      summary: 'CALCULATED -> REVIEWED (§76)'
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Reviewed }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/billing/payouts/{id}/approve:
    post:
      operationId: approvePayout
      tags: [Billing]
      summary: 'REVIEWED or ADJUSTED -> APPROVED (§76, §83.1)'
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Approved }
        '403': { description: 'Requires payout:approve (§66 FINANCE)' }

        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/billing/payouts/{id}/mark-paid:
    post:
      operationId: markPayoutPaid
      tags: [Billing]
      summary: 'APPROVED -> PAID (§76)'
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      responses:
        '201': { description: Paid }
        '409': { description: 'Invalid transition -- a DISPUTED payout cannot be paid (§83.1)' }

        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/billing/payouts/{id}/dispute:
    post:
      operationId: disputePayout
      tags: [Billing]
      summary: Open a payout dispute
      description: |
        §83.1: either side may dispute, and the payout is **held** until the
        dispute resolves.
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Disputed; payout on hold }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/billing/payouts/{id}/resolve-dispute:
    post:
      operationId: resolvePayoutDispute
      tags: [Billing]
      summary: Resolve a dispute
      description: |
        §83.1: "Corrections create adjustment FinancialEvents; never mutate
        historical FinancialEvents." An `ADJUSTMENT` event is **appended**; the
        original accrual is left exactly as it was.
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Resolved }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/billing/invoices/preview/{campaignId}:
    get:
      operationId: previewInvoice
      tags: [Billing]
      summary: Buyer invoice preview
      description: |
        §19. Computed from immutable FinancialEvents. Tax treatment is
        configured per jurisdiction and is **not** assumed here (§81).
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - { name: campaignId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Invoice preview }

        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  # ---------------------------------------------------------------------------
  # ---------------------------------------------------------------------------
  # External channel activation (§15, §16, §17, §47, §48)
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        §64: OIDC access token. Identity is delegated to the provider; Oolix
        stores only the subject claim. Roles are read from the database for the
        active organization, never from a token claim (§4.2).
    agentBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '§92.3: Oolix-issued Agent access token, 15-minute lifetime.'
    crmApiKey:
      type: http
      scheme: bearer
      description: "§71: the Buyer's CRM API key. Hashed at rest and rotatable."
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, maxLength: 200 }
      description: |
        §22.3, §99. Same key + same payload replays the stored result; same key
        + different payload is `409 IDEMPOTENCY_CONFLICT`. **Required** on
        campaign submit, approval and mark-paid (§53).
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, maxLength: 200 }
      description: |
        §53 makes the key MANDATORY here. These three operations each have an
        effect that must not happen twice -- starting a second Partner review,
        creating a second set of activations, or paying a payout twice -- so a
        retry without a key is refused rather than guessed at.
...
    OrgId:
      name: X-Org-Id
      in: header
      required: false
      schema: { type: string, format: uuid }
      description: |
        Selects which organization the request acts within, for a user who
        belongs to several (§34 role switching without a second account).
...
  responses:
    Unauthorized:
      description: '`AUTH_001` -- invalid or expired token.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    RateLimited:
      description: |
        `SYS_001` -- §86's per-class budget for this caller is exhausted. The
        response carries `X-RateLimit-*`, `Retry-After` and
        `retry_after_seconds`, so a client can pace itself instead of guessing
        (§94).
      headers:
        X-RateLimit-Limit:
          { schema: { type: integer }, description: Requests allowed in the window. }
        X-RateLimit-Remaining:
          { schema: { type: integer }, description: Requests left in the window. }
        X-RateLimit-Reset:
          { schema: { type: integer }, description: Unix seconds when the window frees up. }
        Retry-After: { schema: { type: integer }, description: Seconds to wait before retrying. }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      description: |
        §53, §99. Every error uses this envelope: a canonical code, a human
        message, an explicit retryable flag, field errors and the correlation
        id. §99 forbids exposing stack traces or secrets.
      properties:
        error:
          type: object
          required: [code, message, retryable, correlation_id, field_errors]
          properties:
            code: { $ref: '#/components/schemas/ErrorCode' }
            message: { type: string }
            retryable: { type: boolean }
            correlation_id: { type: string }
            retry_after_seconds: { type: integer }
            field_errors:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
    ErrorCode:
      type: string
      description: '§77.2 canonical error registry.'
      enum:
        [
          AUTH_001,
          PERM_001,
          PERM_002,
          VAL_001,
          CAMP_001,
          CAMP_002,
          CAMP_003,
          PART_001,
          PART_002,
          CHAN_001,
          CHAN_002,
          AGENT_001,
          SYS_001,
          SYS_002,
          IDEMPOTENCY_CONFLICT,
        ]

    # --- §89 canonical channel enum ------------------------------------------
...
    LeadEvent:
      type: object
      required: [click_token, crm_event_id, status, event_time]
      properties:
        click_token:
          type: string
          description: '§90: the opaque token the Buyer stored with the lead. Resolved by SHA-256.'
        crm_event_id:
          type: string
          description: '§22.3: idempotency is on `(buyer_org_id, crm_event_id)`.'
        status: { type: string, enum: [RECEIVED, VALID, QUALIFIED, CONVERTED, REJECTED] }
        event_time: { type: string, format: date-time }
        lead_reference: { type: string, description: "The Buyer's own reference. Opaque to Oolix." }
        metadata:
          type: object
          additionalProperties: true
          properties:
            qualification_reason: { type: string }
            rejection_reason: { type: string }
...
    CalculatePayout:
      type: object
      required: [activation_id, period_start, period_end]
      properties:
        activation_id: { type: string, format: uuid }
        period_start: { type: string, format: date-time }
        period_end: { type: string, format: date-time }
    Settlement:
      type: object
      description: |
        §102 worked example: CPQL INR 4,500 x 100 qualified = INR 450,000 media;
        10% platform fee = INR 45,000; Buyer subtotal INR 495,000; Partner
        payout basis INR 450,000.
      properties:
        payout_id: { type: string, format: uuid }
        status: { type: string }
        source_ref: { type: string }
        settlement_inputs:
          type: object
          additionalProperties: true
          description: '§50 rule 4: the immutable inputs the figure was derived from.'
        settlement:
          type: object
          properties:
            payable_count: { type: integer }
            media_amount_minor: { type: integer, format: int64 }
            platform_fee_minor: { type: integer, format: int64 }
            invoice_subtotal_minor: { type: integer, format: int64 }
            partner_payout_basis_minor: { type: integer, format: int64 }
            currency: { type: string }
            settles_on_verified_outcomes: { type: boolean }
...